+++ src/resources/views/layouts/master.blade.php
@@ -25,28 +25,33 @@
                 <div class="container-fluid">
                     <div class="navbar-header">
                         <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-collapse">
-                            <span class="icon-bar"></span>
-                            <span class="icon-bar"></span>
-                            <span class="icon-bar"></span>
+                            <span class="glyphicon glyphicon-user"></span>
                         </button>
-                        <a href="#" class="navbar-brand sidebar-toggle hidden-xs">
-                            <span class="glyphicon glyphicon-menu-hamburger"></span>
+                        <a href="#" class="navbar-brand sidebar-toggle">
+                            <span class="glyphicon glyphicon-option-vertical"></span>
                         </a>
-                        <a href="{{ URL::to('admin') }}" class="navbar-brand visible-xs"><img src="{{ URL::to('vendor/redooor/redminportal/img/favicon.png') }}" title="RedminPortal" class="redooor-nav-logo"> RedminPortal</a>
                     </div>
+                    <ul class="nav navbar-nav navbar-breadcrumb hidden-xs">
+                        <li><a href="{{ URL::to('admin') }}">{{ Lang::get('redminportal::menus.dashboard') }}</a></li>
+                        @section('navbar-breadcrumb')
+                        @show
+                    </ul>
+                    <ul class="nav navbar-nav navbar-right hidden-xs">
+                        <li class="dropdown">
+                            <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false"><span class="glyphicon glyphicon-user"></span>&nbsp;&nbsp; {{ Auth::user()->first_name }} {{ Auth::user()->last_name }}</a>
+                            <ul class="dropdown-menu">
+                                <li><a href="{{ url('myaccount') }}"><span class="glyphicon glyphicon-cog"></span> {{ Lang::get('redminportal::menus.my_account') }}</a></li>
+                                <li><a href="{{ URL::to('logout') }}" title="Lang::get('redminportal::menus.logout')"><span class="glyphicon glyphicon-log-out"></span> {{ Lang::get('redminportal::menus.logout') }}</a></li>
+                            </ul>
+                        </li>
+                    </ul>
                     <div class="navbar-collapse collapse">
-                        <ul class="nav navbar-nav navbar-breadcrumb hidden-xs">
-                            <li><a href="{{ URL::to('admin') }}">{{ Lang::get('redminportal::menus.dashboard') }}</a></li>
-                            @section('navbar-breadcrumb')
-                            @show
-                        </ul>
-                        {{ \Redooor\Redminportal\App\Helpers\RHelper::printMenu(config('redminportal::menu'), 'nav navbar-nav navbar-xs hidden-lg hidden-md hidden-sm') }}
-                        <ul class="nav navbar-nav navbar-right">
-                            <li><a class="btn btn-link hidden-xs" href="{{ URL::to('logout') }}" title="Lang::get('redminportal::menus.logout')">{{ Lang::get('redminportal::menus.logout') }} <i class="glyphicon glyphicon-log-out"></i></a></li>
-                            <li><a class="visible-xs" href="{{ URL::to('logout') }}" title="Lang::get('redminportal::menus.logout')">{{ Lang::get('redminportal::menus.logout') }}</a></li>
+                        <ul class="nav navbar-nav navbar-right visible-xs">
+                            <li class="disabled"><a>Signed in as {{ Auth::user()->first_name }} {{ Auth::user()->last_name }}</a></li>
+                            <li><a href="#"><span class="glyphicon glyphicon-cog"></span> {{ Lang::get('redminportal::menus.my_account') }}</a></li>
+                            <li><a href="{{ URL::to('logout') }}" title="Lang::get('redminportal::menus.logout')"><span class="glyphicon glyphicon-log-out"></span> {{ Lang::get('redminportal::menus.logout') }}</a></li>
                         </ul>
                     </div><!--/.nav-collapse -->
-
                 </div>
             </div>
         </header>